fix install_requires
since the original is subclassing setuptools.commands.install.install it is made impossible to use the install_requires cfg key as intended. This is fixing this problem.

setup.py is now only processing the value replacement with setup() is doing the heavy lifting. The class isn't required to manage that.